refactor(xstateMachineUtils): rename to allStepsFromMachine, move to test folder
